feat(Db): Add SnowflakeId generator
Use generator from godruoyi/php-snowflake adapted for our needs. With a
IMemCache as sequence generator when available or a random sequence
generator otherwise.

Implemented for the Preview table where this allow to save the preview
on the filesystem as we know the id now before inserting it to the DB.

Signed-off-by: Carl Schwan <carl.schwan@nextcloud.com>